Ignore type errors related to verify_match_format
verify_match_format did not return a value, despite the return value having an
important role in the AbbreviationDetector. We decided that preserving the
existing behaviour was probably wiser for now than removing all the relevant code
or reinstating the apparent return value with no real testing.

This probably leads to the AbbrievationDetector finding no abbreviations.