Force enrichment to identify namespace
A namespace can be defined in XML in either the node that uses it or any parent node.
If the `uk` namespace is defined in every node it is used in, rather than the root node,
then adding a new `uk:tna-enrichment-engine` node leaves that namespace undefined,
so we must define it ourselves.